refactor: add folder structure and move files accordingly for user-service.
