+++ user-service/src/entities/User.ts
